server/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
certifi==2024.8.30; python_version >= '3.6'
cffi==1.17.1; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
cryptography==43.0.3; python_version >= '3.7'
defusedxml==0.8.0rc2; python_version >= '3.6'
django==5.1.3; python_version >= '3.10'
django-cors-headers==4.6.0; python_version >= '3.9'
django-redis==5.4.0; python_version >= '3.6'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
djoser==2.3.1; python_version >= '3.8' and python_version < '4.0'
gunicorn==23.0.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
oauthlib==3.2.2; python_version >= '3.6'
packaging==24.2; python_version >= '3.8'
pdfminer.six==20231228; python_version >= '3.6'
pdfplumber==0.11.4; python_version >= '3.8'
pillow==11.0.0; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.9.0; python_version >= '3.8'
pypdf2==3.0.1; python_version >= '3.6'
pypdfium2==4.30.0; python_version >= '3.6'
python3-openid==3.2.0
redis==5.2.0; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
social-auth-app-django==5.4.2; python_version >= '3.8'
social-auth-core==4.5.4; python_version >= '3.8'
sqlparse==0.5.1; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.8'
whitenoise==6.0.0
django-environ

